fixed Font class so that it will use the default VAO. (Had to change order - need to load shaders while VAO is bound; then load meshes (which will unbind the VAO - but the VAO gets reset in draw loop)). May want to make it so that I can pass in a VAO if anything gets complicated... Also had to make sure to use LINEAR filtering (changed default to NEAREST...)

+++ src/Utils.cpp
@@ -26,7 +26,7 @@ namespace aluminum {
 
   float Utils::random() {
     //return glm::compRand1(0,1);
-    return glm::linearRand(0,1);
+    return glm::linearRand(0.0,1.0);
   }
 
   
@@ -41,7 +41,7 @@ namespace aluminum {
     }
 
     //return (long) glm::compRand1(min, max);
-    return (long) glm::linearRand(min, max);
+    return (long) glm::linearRand((double)min, (double)max);
     // return min + int( ((max-min)+1) * random() ); 
   }
 
@@ -55,7 +55,7 @@ namespace aluminum {
       max = max+1;
     }
 
-    return (int) glm::linearRand(min, max);
+    return (int) glm::linearRand((double)min, (double)max);
     //return (int) glm::compRand1(min, max);
     // return min + int( ((max-min)+1) * random() ); 
   }
